requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.2
django-filter==21.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
importlib-metadata==4.11.1
Markdown==3.3.6
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
zipp==3.7.0
